--- Jonathan's_Info/static/css/style.css
@@ -1,236 +1,0 @@
-.active {
-  opacity: 1;
-}
-
-.inactive {
-  opacity: 0.5;
-}
-
-.table-success,.table-success>td,.table-success>th{
-background-color:white}
-
-.table-success tbody+tbody,.table-success td,.table-success th,.table-success thead th{
-border-color:#a1deb6}
-
-.table-hover .table-success:hover{
-background-color:#bae6c9}
-
-.table-hover .table-success:hover>td,.table-hover .table-success:hover>th{
-background-color:#bae6c9}
-
-.table-success,.table-success>td,.table-success>th{
-background-color:#1ED760
-}
-
-.line {
-  fill: none;
-  stroke-width: 2px;
-}
-
-.green {
-  stroke: #1ED760;
-}
-
-.mintgreen {
-  stroke: #c3f0c8
-}
-
-.purple {
-  stroke: #af2896;
-}
-
-.blue {
-  stroke: #2d46b9;
-}
-
-.lightblue {
-  stroke: #007aff
-}
-
-.yellow {
-  stroke: #ffc864;
-}
-
-.pink {
-  stroke: #f573a0;
-}
-
-.red {
-  stroke: #eb1e32;
-}
-
-.legend {
-  height: 300px; 
-  float: right;
-}
-
-.right {
-float: right;
-}
-
-.ufo-form {
-margin-top: 50px;
-}
-*/
-.filters .panel-body {
-padding: 0;
-}
-
-.filters {
-margin-top: 10px;
-}
-*/
-select {
-display: block;
-width: 100%;
-padding: 8px;
-color: #8e979f;
-outline: none;
-}
-
-#table-area th {
-text-align: center;
-}
-
-#table-area th:first-child,
-#table-area th:nth-child(2),
-#table-area th:nth-child(6) {
-min-width: 70px;
-}
-
-.loader {
-margin-top: 50px;
-}
-
-.margin-top-50 {
-margin-top: 50px;
-}
-
-/* input placeholder color changes */
-.form-control::-webkit-input-placeholder {
-/* Chrome/Opera/Safari */
-color: #8e979f;
-}
-
-.form-control::-moz-placeholder {
-/* Firefox 19+ */
-color: #8e979f;
-}
-
-.form-control:-ms-input-placeholder {
-/* IE 10+ */
-color: #8e979f;
-}
-
-.form-control:-moz-placeholder {
-/* Firefox 18- */
-color: #8e979f;
-}
-
-/* media queries which will only be activated at specific screen sizes */
-@media (max-width: 800px) {
-#table td {
-  font-size: 1em;
-}
-}
-
-@media (max-width: 500px) {
-#table thead {
-  display: none;
-}
-
-#table tr td:first-child {
-  font-size: 1.8em;
-  font-weight: bold;
-  background: #1b3146;
-}
-
-#table tr td:last-child {
-  margin-bottom: 20px;
-}
-
-#table tr td:last-child::after {
-  display: block;
-  width: 100%;
-  height: 1px;
-  margin-top: 20px;
-  content: " ";
-  background: aliceblue;
-}
-
-#table td {
-  font-size: 1.25em;
-}
-
-#table tbody td {
-  display: block;
-  text-align: center;
-}
-
-#table tbody td::before {
-  display: block;
-  font-size: 1.4em;
-  font-weight: bold;
-  text-align: center;
-  content: attr(data-th);
-}
-}
-
-.center {
-display: block;
-margin-left: auto;
-margin-right: auto;
-width: 50%;
-}
-
-.modal {
-display: none;
-position: fixed; 
-padding-top: 50px;
-left: 0; 
-top: 0;
-width: 100%;
-height: 100%; 
-background-color: rgb(0, 0, 0);
-background-color: rgba(0, 0, 0, 0.5);
-}
-.modal-content {
-position: relative; 
-background-color: white;
-padding: 20px; 
-margin: auto; 
-width: 75%;  
--webkit-animation-name: animatetop;
--webkit-animation-duration: 0.4s;
-animation-name: animatetop;
-animation-duration: 0.4s
-}
-.close-btn {
-float: right; 
-color: lightgray; 
-font-size: 24px;  
-font-weight: bold;
-}
-.close-btn:hover {
-color: darkgray;
-}
-@-webkit-keyframes animatetop {
-from {top:-300px; opacity:0} 
-to {top:0; opacity:1}
-}
-@keyframes animatetop {
-from {top:-300px; opacity:0}
-to {top:0; opacity:1}
-}
-
-/* .jumbotron {
-background-image: url("/static/images/skyline.png");
-background-size: 100%;
-} */
-
-.title-center {
-position: absolute;
-top: 100%;
-left: 50%;
-transform: translate(-50%, -50%);
-}
