--- Jonathan's_Info/templates/about.html
@@ -1,143 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <title>About Us</title>
-
-    <!-- Bootstrap and Bootswatch CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootswatch/4.5.2/cosmo/bootstrap.min.css">
-    <link rel="stylesheet" type="text/css" href="static/css/style.css">
-    <!--<link rel="stylesheet" href="assets/css/d3Style.css">-->
-
-   <!-- <link rel="stylesheet"  href="style.css">-->
-    <script src="https://d3js.org/d3.v6.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/d3-legend/2.25.6/d3-legend.min.js"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-    <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/d3-tip/0.7.1/d3-tip.min.js"></script> -->
-    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-    <script src="https://unpkg.com/d3-v6-tip@1.0.6/build/d3-v6-tip.js"></script>
-    <script src="https://unpkg.com/jeezy@1.13.1/lib/jeezy.min.js"></script>
-    <!--<script type="text/javascript" src="static/js/app.js"></script>-->
-</head>
-
-<body>
-
-    <!--Navbar-->
-    <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
-        <a class="navbar-brand" href="/">
-            <img src="../static/images/dallas_logo.jpeg" alt="logo" style="width:40px;height:40px;">
-        </a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarColor02" aria-controls="navbarColor02" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-        </button>
-
-        <div class="collapse navbar-collapse" id="navbarColor01">
-            <ul class="navbar-nav ml-auto">
-                <li class="nav-item">
-                    <a class="nav-link" href="/">Home</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/dashboard">Visualizations</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/ml">Status Predictor</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/exploration">Crime Analysis</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/data">Data</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="/resources">Resources</a>
-                </li>
-                <li class="nav-item active">
-                    <a class="nav-link" href="/about">About</a>
-                </li>
-            </ul>
-        </div>
-        
-    </nav>
-
-    <!--JUMBOTRON-->
-   <!--<div class="jumbotron text-center">
-        <h3 class="display-3">About Us</h3>
-    </div>-->
-
-    <div class="row text-center" style="margin-top:3%">
-        <div class = "col-md-12">
-            <h6 class="display-4">About Us</h6>
-        </div>
-    </div>
-
-    <!--ABOUT-->
-    <div class= "container-fluid">
-        <div class="row" style="margin-top:5%">
-            <div class = "col-md-12">
-                <h5>Alex Arnold</h5>
-                <br>
-            </div>
-        </div>
-        <div class="row" >
-            <div class = "col-md-3">
-                <img src="../static/images/alex.JPG" alt="Alex" style="width:65%;padding:2%;" class="center">
-            </div>
-            <div class = "col-md-9">
-                <blockquote class="blockquote">
-                    <p class="mb-0">Alex Arnold graduated from Santa Clara University with a Bachelors of Science in Marketing and Psychology.  
-                        She has over five years of professional leadership experience in retail. After spending the last few years 
-                        of her career in the role of mom, she is expanding her education into the realm of analytics and data 
-                        science as she has found her passion for problem solving and turning data into insights. She enjoys 
-                        traveling and spending time with her three children.</p>
-                  </blockquote>
-            </div>
-        </div>
-        <hr>
-        <div class="row" style="margin-top:5%">
-            <div class = "col-md-12">
-                <h5>Samantha Lane</h5>
-                <br>
-            </div>
-        </div>
-        <div class="row" >
-            <div class = "col-md-9">
-                <blockquote class="blockquote">
-                    <p class="mb-0">Samantha Lane is a licensed, registered dietitian with a Masters of Science specializing in Nutrition
-                        and Metabolism from the University of Texas Medical Branch. She has worked as a registered dietitian in a variety of 
-                        setting over the last 2 years. As she works on furthering her education, she has developed a passion for diving into 
-                        big data and developing actionable insights. Her hobbies include playing with her English Bulldog, Ollie, and finding 
-                        new recipes to enhance her cooking abilities.</p>
-                </blockquote>
-            </div>
-            <div class = "col-md-3">
-                <img src="../static/images/sam.jpg" alt="Sam" style="width:65%;padding:2%;" class="center">
-            </div>
-        </div>
-        <hr>
-        <div class="row" style="margin-top:5%">
-            <div class = "col-md-12">
-                <h5>Rachel Podemski</h5>
-                <br>
-            </div>
-        </div>
-        <div class="row" >
-            <div class = "col-md-3">
-                <img src="../static/images/rachel.jpg" alt="Rachel" style="width:65%;padding:2%;" class="center">
-            </div>
-            <div class = "col-md-9">
-                <blockquote class="blockquote">
-                    <p class="mb-0">Rachel Podemski graduated from Texas Christian University with a 
-                        Bachelors of Science in Economics. She has worked at BSN Sports for over three 
-                        years now, and has developed a passion to work with big data and sports. 
-                        She enjoys spending time with her three dogs and traveling.</p>
-                </blockquote>
-            </div>
-        </div>
-        <hr>
-    </div>
-
-</body>
-
-</html>
